+++ docs/source/conf.py
@@ -22,6 +22,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.intersphinx',
+    'myst_nb',
 ]
 
 intersphinx_mapping = {
