+++ docs/source/index.rst
@@ -1,4 +1,8 @@
 Metasmith
 ===================================
 
-placeholder
+.. toctree::
+   :maxdepth: 2
+   :caption: asdf
+
+   minimal_example
