+++ envs/docs.yml
@@ -4,3 +4,5 @@ dependencies:
   - python=3.12
   - sphinx=8.2.3
   - sphinx-rtd-theme=3.0.1
+  - sphinx-autobuild=2024.10.3
+  - myst-nb=1.2.0
