+++ main/local_mock/transforms/dupe_test/diamond.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from metasmith.pythonapi import *
+from metasmith.python_api import *
 
 # todo: url for more consistency
 lib = DataTypeLibrary.Load("/home/tony/workspace/tools/Metasmith/main/local_mock/prototypes/metagenomics.dev3.yml")
