+++ src/metasmith/version.txt
@@ -1,1 +1,1 @@
-0.5
+0.5.1
